+++ config/app.php
@@ -15,7 +15,7 @@ return [
 
     'name' => env('APP_NAME', 'V-Air Ops'),
 
-    'version' => env('APP_VERSION', 'v1.1.9'),
+    'version' => env('APP_VERSION', 'v1.1.10'),
 
     /*
     |--------------------------------------------------------------------------